# Refactor `LinkParser`