--- a/src/common/LinkParser.cpp
+++ b/src/common/LinkParser.cpp
@@ -19,7 +19,7 @@ QSet<QString> &tlds()
 #if QT_VERSION >= QT_VERSION_CHECK(6, 0, 0)
         // Default encoding of QTextStream is already UTF-8, at least in Qt6
 #else
-        stream.setCodec(QStringLiteral("UTF-8"));
+        stream.setCodec("UTF-8");
 #endif
         int safetyMax = 20000;