--- a/src/common/LinkParser.cpp
+++ b/src/common/LinkParser.cpp
@@ -123,6 +123,15 @@ LinkParser::LinkParser(const QString &unparsedString)
     QStringView remaining(unparsedString);
     QStringView protocol(remaining);
 
+#if QT_VERSION < QT_VERSION_CHECK(5, 15, 0)
+    QStringView wholeString(unparsedString);
+    const auto refFromView = [&](QStringView view) {
+        return QStringRef(&unparsedString,
+                          static_cast<int>(view.begin() - wholeString.begin()),
+                          static_cast<int>(view.size()));
+    };
+#endif
+
     // Check protocol for https?://
     if (remaining.startsWith(QStringLiteral("http"), Qt::CaseInsensitive) &&
         remaining.length() >= 4 + 3 + 1)  // 'http' + '://' + [any]
@@ -137,7 +146,12 @@ LinkParser::LinkParser(const QString &unparsedString)
         if (remaining.startsWith(QStringLiteral("://")))
         {
             remaining = remaining.mid(3);
+#if QT_VERSION >= QT_VERSION_CHECK(5, 15, 0)
             result.protocol = {protocol.begin(), remaining.begin()};
+#else
+            result.protocol =
+                refFromView({protocol.begin(), remaining.begin()});
+#endif
         }
     }
 
@@ -202,8 +216,13 @@ LinkParser::LinkParser(const QString &unparsedString)
     if ((nDots == 3 && isValidIpv4(host)) ||
         isValidTld(host.mid(lastDotPos + 1)))
     {
+#if QT_VERSION >= QT_VERSION_CHECK(5, 15, 0)
         result.host = host;
         result.rest = rest;
+#else
+        result.host = refFromView(host);
+        result.rest = refFromView(rest);
+#endif
         result.source = unparsedString;
         this->result_ = std::move(result);
     }