# Description

This PR is the continuation of #4507 and #4436.

These are the main changes:

- Use custom parsing for IPv4 addresses over a regex.
- Drop IPv6 support (was disabled by default).
- Use `QStringView` over `QStringRef`
- Removed usage of `goto`.

In my tests (`chatterino-benchmark --benchmark_filter=LinkParsing --benchmark_repetitions=100`), this version is about 6x faster:

### Before

| Benchmark | Time| CPU | Iterations|
|--------|--------|--------|--------|
|BM_LinkParsing    |         33159 ns    |    33692 ns   |     21333 |
|... |
|BM_LinkParsing_mean    |    33569 ns     |   33523 ns         | 100 |
|BM_LinkParsing_median  |    33178 ns     |   32959 ns         | 100 |
|BM_LinkParsing_stddev    |   1520 ns     |    1457 ns         | 100 |
|BM_LinkParsing_cv       |    4.53 %      |    4.35 %          | 100 |

![flamegraph](https://user-images.githubusercontent.com/19953266/233839941-93c45e6d-b3d7-4ae7-bf42-a47cceaef1c3.png)

### After
| Benchmark | Time| CPU | Iterations|
|--------|--------|--------|--------|
|BM_LinkParsing            |  5137 ns      |   5022 ns     |  112000|
|...|
|BM_LinkParsing_mean       |  5369 ns      |   5352 ns     |     100|
|BM_LinkParsing_median     |  5324 ns      |   5301 ns     |     100|
|BM_LinkParsing_stddev     |   273 ns      |    254 ns     |     100|
|BM_LinkParsing_cv         |  5.08 %       |   4.75 %      |     100|

![flamegraph](https://user-images.githubusercontent.com/19953266/233839964-526c7db1-2517-4750-8fd3-01cfa1c5b1df.png)

As you can see, I didn't change the `isValidTld` it's now the bottleneck here. The main reason is that a `QString` needs to be created and case-converted. Ideally, it could use a `QStringView` and have case-insensitive search functions.