@pajlada 2023-04-23
Code looks good, some small nitpicks that you can change if you feel like it's reasonable